+++ tests/model/test_model.py
@@ -12,6 +12,10 @@ class TestYacgModel (unittest.TestCase):
         x = yacg.model.model.Type()
         self.assertIsNotNone(x)
 
+    def testObjectType(self):
+        x = yacg.model.model.ObjectType()
+        self.assertIsNotNone(x)
+
     def testIntegerType(self):
         x = yacg.model.model.IntegerType()
         self.assertIsNotNone(x)