+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 3.2.0
-* implement model support for dictionaries 
+* implement model support for dictionaries
+* introduce pure object type
 
 # 3.1.0
 * improved Python object initialization - inspired by a contribution of Paul Way [https://github.com/PaulWay]